Use face cache stored in font when possible
This commit adds a font face cache slot to the font class which gets used when
possible. If it is for some reason not available then the default cache is
used. This is made available by binding *current-font* to the font in question
before rendering something, ie in a let form.

Additionally an optional overwrite argument was added to the function
cache-char, which when true replaces the rendered char in the hash table with a
newly rendered char.